Feat: homework is done

+++ src/main/java/com/dmitrySheyko/app/services/BookService.java
@@ -1,20 +1,20 @@
-package com.dmitrySheyko.app;
+package com.dmitrySheyko.app.services;
 
-import com.dmitrySheyko.app.services.ProjectRepository;
+import com.dmitrySheyko.app.repositories.BookRepository;
+import com.dmitrySheyko.app.repositories.ProjectRepository;
 import com.dmitrySheyko.web.dto.Book;
+import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.thymeleaf.util.StringUtils;
 
 import java.util.List;
-
-import static org.thymeleaf.util.StringUtils.*;
 
 @Service
 public class BookService {
 
     private final ProjectRepository<Book> bookRepo;
+    private final Logger logger = Logger.getLogger(BookService.class);
 
     @Autowired
     public BookService(BookRepository<Book> bookRepo) {
@@ -26,13 +26,34 @@ public class BookService {
     }
 
     public void save(Book book) {
+        logger.info("Attempt of saving book in bookService");
         if (isAuthorCorrect(book.getAuthor()) || isTitleCorrect(book.getTitle()) || isSizeCorrect(book.getSize())) {
             bookRepo.store(book);
+        } else {
+            logger.info("Book hasn't saved. Book description is incorrect");
         }
     }
 
     public void removeBookId(Integer bookIdToRemove) {
-        bookRepo.removeItemById(bookIdToRemove);
+        logger.info("Attempt of removing book by id in bookService");
+        if (isBookIdCorrect(bookIdToRemove)) {
+            bookRepo.removeItemById(bookIdToRemove);
+        } else {
+            logger.info("Book hasn't removed. Book id is incorrect");
+        }
+    }
+
+    public void removeBooksByRegex(String regex) {
+        logger.info("Attempt of removing book by regex in bookService");
+        if (isRegexCorrect(regex)) {
+            String[] searchParameters = regex.split("/");
+            String authorForDelete = searchParameters[0];
+            String titleForDelete = searchParameters[1];
+            String sizeForDelete = searchParameters[2];
+            bookRepo.removeByRegex(authorForDelete, titleForDelete, sizeForDelete);
+        } else {
+            logger.info("Books hasn't removed. Regex is incorrect");
+        }
     }
 
     private boolean isAuthorCorrect(String author) {
@@ -44,19 +65,11 @@ public class BookService {
     }
 
     private boolean isSizeCorrect(Integer size) {
-        if (size == null) {
-            return false;
-        }
-        if (size < 0) {
-            return false;
-        }
-        return true;
+        return size != null && size >= 0;
     }
 
-    public void removeBookByRegex(String regex) {
-        String[] searchParameters = regex.split("/");
-        bookRepo.removeByRegex(searchParameters);
-
+    private boolean isBookIdCorrect(Integer bookId) {
+        return (bookId != null && bookId > 0);
     }
 
     private boolean isRegexCorrect(String regex) {
